refactor show rock climb details function so it does not require Int64 type declaration

+++ BMG/ViewModel/LogbookViewModel.swift
@@ -30,7 +30,7 @@ class LogbookViewModel {
     var idOfRockClimbSelected: Int64 = -1
     
     func showRockClimbDetails(){
-        screenNavigationController.displayLoggedRockClimbDataScreen(Int64(idOfRockClimbSelected))
+        screenNavigationController.displayLoggedRockClimbDataScreen(idOfRockClimbSelected)
     }
     
 }
